deleted synthetic module, deleted test/integration/synthetic_generation/, moved synthetic/data.py one level up
