+++ examples/ts_gapfilling_example.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from fedot.core.chains.chain import Chain
 from fedot.core.chains.node import PrimaryNode, SecondaryNode
-from fedot.utilities.synthetic.data import generate_synthetic_data
+from fedot.utilities.data import generate_synthetic_data
 from fedot.utilities.ts_gapfilling import ModelGapFiller, SimpleGapFiller
