+++ fedot/utilities/data.py
